Move toasts folder into team folder and rename it
